// interfaces/AppInterfaces.tsx
// Generated by https://quicktype.io
export interface LoginData{
    correo: string;
    contra: string;
}

export interface LoginResponse {
    correo: Usuario;
    token:   string;
    mensaje?: string;
}

export interface Usuario {
    userid: string;
    nombre: string;
    apellido: string;
    correo: string;
}

export interface infoUsuario {
    userid: string;
    nombre: string;
    apellido: string;
    correo: string;
}

export interface ListadoAlumnos{
    id: string;
    nombre: string;
    apellido: string;
}

// Generated by https://quicktype.io
// Generated by https://quicktype.io

// Generated by https://quicktype.io



export interface AlumnoResponse {
    total:  number;
    alumno: Alumno[];
}

export interface Alumno {
    ide: string;
    abre: string;
    nombredoc: string;
    nombre: string;
    apellido: string;
    fechar: string;
    fechan: string;
    lnacimiento: string;
    genero: string;
    colegio: string;
    estado: string;
    nacionalidad: string;
    dir_estado: string;
    dir_municipio: string;
    dir_localidad: string;
    direccion: string;
    cp: string;
    celular: string;
    correo: string;
    imagen: string;
}